+++ src/Narvalo.Core/Fx/Either`2.cs
@@ -118,6 +118,8 @@ namespace Narvalo.Fx
 
             public override string ToString()
             {
+                Contract.Ensures(Contract.Result<string>() != null);
+
                 return Format.CurrentCulture("Left({0})", _value);
             }
         }
@@ -182,6 +184,8 @@ namespace Narvalo.Fx
 
             public override string ToString()
             {
+                Contract.Ensures(Contract.Result<string>() != null);
+
                 return Format.CurrentCulture("Right({0})", _value);
             }
         }
@@ -221,13 +225,6 @@ namespace Narvalo.Fx
             Contract.Ensures(Contract.Result<Maybe<TRight>>() != null);
 
             return default(Maybe<TRight>);
-        }
-
-        public override string ToString()
-        {
-            Contract.Ensures(Contract.Result<string>() != null);
-
-            return String.Empty;
         }
     }